listing page - cleanup old variables, use resort-rules component, rename footer and fix bg
