+++ src/components/Home.vue
@@ -198,17 +198,17 @@
         </v-flex>
       </v-row>
     </v-container>
-    <Footer></Footer>
+    <page-footer></page-footer>
   </div>
 </template>
 
 <script>
 import { PageService } from "@/connection/resources.js"
-const Footer = () => import ('@/components/Footer.vue')
+const PageFooter = () => import ('@/components/PageFooter.vue')
 
 export default {
   components:{
-    Footer,
+    PageFooter,
   },
   data() {
     return {
