+++ src/components/PageFooter.vue
@@ -1,15 +1,11 @@
 <template>
-  <v-footer v-resize="storeFooterHeight" dark padless class="page-footer" ref="pageFooter">
-    <v-container
-      fluid
-      class="px-0"
-      style="box-shadow: 0px -1px 0px #121416; color: #D8DADE;"
-    >
+  <v-footer v-resize="storeFooterHeight" padless class="page-footer" ref="pageFooter">
+    <v-container fluid class="px-0" style="box-shadow: 0px -1px 0px #121416; color: #D8DADE;">
       <v-container class="is-limited align-start" id="footer">
         <v-row no-gutters justify-space-between>
           <v-col cols="6" sm="4" class="mb-4 mb-sm-0">
             <h3 class="h3 mb-8">Engage</h3>
-            
+
             <ul style="list-style:none; padding:0; margin:0;">
               <li class="mb-2">
                 <router-link to="/Contact" class="link">
@@ -17,9 +13,7 @@
                 </router-link>
               </li>
               <li class="mb-2">
-                <a href="http://asiato.asia/" target="_blank" class="link"
-                  >About</a
-                >
+                <a href="http://asiato.asia/" target="_blank" class="link">About</a>
               </li>
               <!-- <li class="mb-2">Book Now</li> -->
               <li class="mb-2">
@@ -31,13 +25,10 @@
           </v-col>
           <v-col cols="6" sm="4" class="mb-4 mb-sm-0">
             <h3 class="h3 mb-8">Invest</h3>
-            
+
             <ul style="list-style:none; padding:0; margin:0;">
               <li class="mb-2">
-                <router-link
-                  to="/listing/nature-city-investment-cambodia-property"
-                  class="link"
-                >
+                <router-link to="/listing/nature-city-investment-cambodia-property" class="link">
                   Lease a Property
                 </router-link>
               </li>
@@ -51,7 +42,7 @@
           </v-col>
           <v-col cols="12" sm="4" class="mb-4 mb-sm-0">
             <h3 class="h3 mb-8">Learn</h3>
-            
+
             <ul style="list-style:none; padding:0; margin:0;">
               <!-- <li class="mb-2">Community Outreach</li> -->
               <li class="mb-2">
@@ -60,10 +51,7 @@
                 </router-link>
               </li>
               <li class="mb-2">
-                <router-link
-                  to="/listing/Proudly-Powered-by-KIT-Students"
-                  class="link"
-                >
+                <router-link to="/listing/Proudly-Powered-by-KIT-Students" class="link">
                   Proudly Powered by KIT
                 </router-link>
               </li>
@@ -78,6 +66,7 @@
 <script>
 import store from '@/store'
 export default {
+  name: 'page-footer',
   mounted() {
     this.storeFooterHeight()
   },
@@ -96,6 +85,6 @@ export default {
   color: #d8dade;
 }
 .page-footer {
-  background-color: $dark;
+  background-color: $dark !important;
 }
 </style>
